third_party/WebKit/LayoutTests/http/tests/security/no-popup-from-sandbox-top-expected.txt

CONSOLE ERROR: line 1: Unsafe JavaScript attempt to initiate navigation for frame with URL 'http://127.0.0.1:8000/security/no-popup-from-sandbox-top.html' from frame with URL 'data:text/html,       <script>       var win = window.open('about:blank', '_top');       console.log(win ? 'FAIL' : 'PASS');       </script>'. The frame attempting navigation of the top-level window is sandboxed, but the flag of 'allow-top-navigation' or 'allow-top-navigation-by-user-activation' is not set.

CONSOLE MESSAGE: line 1: PASS
To run this test outside of DumpRenderTree, please disable your popup blocker!
